+++ common/src/main/java/traben/entity_model_features/config/EMFConfig.java
@@ -63,6 +63,7 @@ public class EMFConfig extends TConfig {
     public boolean onlyDebugRenderOnHover = false;
     public boolean enforceOptifineSubFoldersVariantOnly = true;
     public boolean enforceOptiFineAnimSyntaxLimits = true;
+    public boolean enforceOptiFineFloorUVs = true;
 
     @Override
     public TConfigEntryCategory getGUIOptions() {
@@ -123,7 +124,9 @@ public class EMFConfig extends TConfig {
                         new TConfigEntryBoolean("entity_model_features.config.optifine_subfolders", "entity_model_features.config.optifine_subfolders.tooltip",
                                 () -> enforceOptifineSubFoldersVariantOnly, value -> enforceOptifineSubFoldersVariantOnly = value, true),
                         new TConfigEntryBoolean("entity_model_features.config.optifine_syntax", "entity_model_features.config.optifine_syntax.tooltip",
-                                () -> enforceOptiFineAnimSyntaxLimits, value -> enforceOptiFineAnimSyntaxLimits = value, true)
+                                () -> enforceOptiFineAnimSyntaxLimits, value -> enforceOptiFineAnimSyntaxLimits = value, true),
+                        new TConfigEntryBoolean("entity_model_features.config.optifine_floor", "entity_model_features.config.optifine_floor.tooltip",
+                                () -> enforceOptiFineFloorUVs, value -> enforceOptiFineFloorUVs = value, true)
 
                 )
         );