+++ common/src/main/java/traben/entity_model_features/models/jem_objects/EMFBoxData.java
@@ -1,5 +1,6 @@
 package traben.entity_model_features.models.jem_objects;
 
+import traben.entity_model_features.EMF;
 import traben.entity_model_features.utils.EMFUtils;
 
 import java.util.Arrays;
@@ -79,17 +80,22 @@ public class EMFBoxData {
         if (uv.length != 4) {
             throw new IllegalArgumentException("Invalid UV data for ["+name+"], must have 4 or 0 values: " + Arrays.toString(uv));
         }
+
         //first two should be integers
-        if (uv[0] != (int) uv[0] || uv[1] != (int) uv[1]) {
-            EMFUtils.logWarn("Possibly invalid UV data for ["+name+"], the first 2 values should be integers (whole numbers), because OptiFine floors them: " + Arrays.toString(uv));
-        }
-        //second two should be 0 or abs() >=1
-        if (uv[2] != 0 && Math.abs(uv[2]) < 1) {
-            EMFUtils.logWarn("Possibly invalid UV data for ["+name+"], the third value should be either 0, less than -1, or larger than 1: " + Arrays.toString(uv));
+        if(EMF.config().getConfig().enforceOptiFineFloorUVs ) {
+            uv[0] = (float) Math.floor(uv[0]);
+            uv[1] = (float) Math.floor(uv[1]);
+            uv[2] = (float) Math.floor(uv[2]);
+            uv[3] = (float) Math.floor(uv[3]);
+        } else if (uv[0] != Math.floor(uv[0]) || uv[1] != Math.floor(uv[1]) || uv[2] != Math.floor(uv[2]) || uv[3] != Math.floor(uv[3])) {
+            EMFUtils.logWarn("Possibly invalid UV data for [" + name + "], all values should be integers (whole numbers), because OptiFine floors them, EMF can be set to floor these values in the OptiFine parity settings: " + Arrays.toString(uv));
         }
 
-        if (uv[3] != 0 && Math.abs(uv[3]) < 1) {
-            EMFUtils.logWarn("Possibly invalid UV data for ["+name+"], the fourth value should be either 0, less than -1, or larger than 1: " + Arrays.toString(uv));
+        if (Math.floor(uv[2]) == 0) {
+            EMFUtils.logWarn("Possibly invalid UV data for ["+name+"], the third value should not floor to 0: " + Arrays.toString(uv));
+        }
+        if (Math.floor(uv[3]) == 0) {
+            EMFUtils.logWarn("Possibly invalid UV data for ["+name+"], the fourth value should not floor to 0: " + Arrays.toString(uv));
         }
 
     }